Create PropertyLoaderClass
This is a basic Class that can load a local file and pull properties from it. From there, you can set variables (Strings s = prop.getProperty(data)). Please note, for this example to work, you have to create a file 'properties/config.properties' and add data in the format of identifier=data, and call the data in the class by the identifier.